backend/requirements.txt
contextlib
fastapi
uvicorn
datetime
typing
pydantic_settings
sqlmodel
email-validator
python-multipart

pytest
